Create shared OAuth2 authorization helpers
- Add oauth2_common module with shared AuthorizeData struct
- Create generic get_authorize_data function
- Refactor Google and Apple modules to use shared logic
- Eliminate ~40 lines of duplicated authorization code
