+++ tee-worker/omni-executor/heima/identity-verification/src/web2/apple/mod.rs
@@ -1,7 +1,7 @@
-use crate::helpers;
 use base64::prelude::{Engine, BASE64_URL_SAFE_NO_PAD};
 use serde::Deserialize;
-use url::Url;
+
+pub use super::oauth2_common::AuthorizeData;
 
 const BASE_URL: &str = "https://appleid.apple.com/auth/authorize";
 const SCOPES: &str = "email";
@@ -17,24 +17,8 @@ pub struct IdToken {
 	pub email_verified: Option<String>,
 }
 
-pub struct AuthorizeData {
-	pub authorize_url: String,
-	pub state: String,
-}
-
 pub fn get_authorize_data(client_id: &str, redirect_uri: &str) -> AuthorizeData {
-	let state = helpers::generate_alphanumeric_otp(32);
-	let mut authorize_url = Url::parse(BASE_URL).expect("Failed to parse URL");
-	authorize_url.query_pairs_mut().extend_pairs(&[
-		("response_type", "code"),
-		("client_id", client_id),
-		("redirect_uri", redirect_uri),
-		("scope", SCOPES),
-		("state", &state),
-		("response_mode", "form_post"),
-	]);
-
-	AuthorizeData { authorize_url: authorize_url.into(), state }
+	super::oauth2_common::get_authorize_data(BASE_URL, client_id, redirect_uri, SCOPES, true)
 }
 
 pub fn decode_id_token(token: &str) -> Result<IdToken, &'static str> {
